+++ packages/all-in/src/lib/graphql/kitQLModules.ts
@@ -1,9 +1,14 @@
 import { useGraphQLModules } from '@envelop/graphql-modules'
 import { createApplication, type Module } from 'graphql-modules'
 
-export const kitQLModules = (modules: Module[]) =>
-  useGraphQLModules(
-    createApplication({
-      modules,
-    })
-  )
+export const kitQLModules = (modules: Module[]) => {
+  try {
+    return useGraphQLModules(
+      createApplication({
+        modules,
+      })
+    )
+  } catch (error) {
+    console.error(error)
+  }
+}