+++ packages/all-in/vite.config.js
@@ -4,9 +4,15 @@ import { sveltekit } from '@sveltejs/kit/vite'
 const config = {
   plugins: [sveltekit()],
 
+  optimizeDeps: {
+    include: ['safe-stable-stringify'],
+  },
+
   build: {
     rollupOptions: {
-      external: '@graphql-yoga/render-graphiql', // Users will decide to opt-in by adding this dep (or not)
+      external: [
+        '@graphql-yoga/render-graphiql', // Users will decide to opt-in by adding this dep (or not)
+      ],
     },
   },
 }