+++ packages/all-in/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kitql/all-in
 
+## 0.8.0-next.0
+
+### Minor Changes
+
+- [#215](https://github.com/jycouet/kitql/pull/215) [`6261921`](https://github.com/jycouet/kitql/commit/6261921157c3581258fa12c1a57fd450f5b26547) Thanks [@jycouet](https://github.com/jycouet)! - update with sveltekit next.483 & import form @kitql/all-in
+
+### Patch Changes
+
+- [#222](https://github.com/jycouet/kitql/pull/222) [`d8a2095`](https://github.com/jycouet/kitql/commit/d8a2095c58c6dd967040c11888d49562c31e66ee) Thanks [@david-plugge](https://github.com/david-plugge)! - Improve GraphiQL offline / online option
+
+- Updated dependencies [[`6261921`](https://github.com/jycouet/kitql/commit/6261921157c3581258fa12c1a57fd450f5b26547)]:
+  - @kitql/helper@0.6.0-next.0
+  - @kitql/module-codegen@0.4.0-next.0
+  - vite-plugin-watch-and-run@1.1.0-next.0
+
 ## 0.7.10
 
 ### Patch Changes